third_party/blink/web_tests/fast/loader/stateobjects/document-destroyed-navigate-back-with-fragment-scroll-expected.txt

ALERT: LOADED
ALERT: State popped - FirstEntry (type string)
ALERT: hashChanged - Last path component of location is document-destroyed-navigate-back-with-fragment-scroll.html#FirstEntry
ALERT: State popped - SecondEntry (type string)
ALERT: hashChanged - Last path component of location is document-destroyed-navigate-back-with-fragment-scroll.html#SecondEntry
ALERT: Navigating back...
ALERT: LOADED
ALERT: Page shown - SecondEntryWillLaterBeReactivated (type string)
ALERT: State popped - FirstEntryWillLaterBeReactivated (type string)
ALERT: hashChanged - Last path component of location is document-destroyed-navigate-back-with-fragment-scroll.html#FirstEntryWillLaterBeReactivated
ALERT: Test complete
This test:
-Builds up a list of state object entries with fragment URLs.
-Navigates through them to verify that the popstate and hashchanged events are fired.
-Navigates away to a new document, with the old document being destroyed.
-Navigates back to the state object entries and verifies the pageshow or popstate events are fired on the new documents.
